[FIX] Add fallback to World Loading (Fixes #29)
When the world preset is null, world dimension creation fails. When re-creating a world, this is the case.
A fallback is introduced that uses the context to create the world dimensions.
This was successfully tested for re-creating worlds that use presets, such as AMPLIFIED, in which case the preset is preserved during preview and generation.